git siff and git diff --cached.
git diff gives you the changes that have been done but have yet to be
added to the index. git diff --cached shows you the changes that will be
committed. Care should be taken that the files that have been changed have
to be staged to commit by adding them to the index using git add.